# backend/src/models/Warehouse.js
const { DataTypes } = require("sequelize");

module.exports = (sequelize) => {
  const Warehouse = sequelize.define(
    "Warehouse",
    {
      id: {
        type: DataTypes.UUID,
        defaultValue: DataTypes.UUIDV4,
        primaryKey: true,
      },
      name: {
        type: DataTypes.STRING,
        allowNull: false,
        unique: true,
      },
    },
    {
      timestamps: true,
      tableName: "warehouses",
    }
  );

  return Warehouse;
};